python_nemo/requirements.txt
asgiref==3.2.3
bson==0.5.8
certifi==2019.11.28
chardet==3.0.4
dataclasses==0.7
Django==3.0.2
djongo==1.3.0
idna==2.8
pymongo==3.10.1
python-dateutil==2.8.1
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.2.4
urllib3==1.25.8

